Add example raw data files from SI Config Plus
The latest version of the SportIdent Config Plus software outputs a CSV file with a different series of fieldnames compared to the legacy version that I currently use. A notable change is the addition of columns for controls up to 192. The additional columns are included without regard for the maximum number of controls punched by any participant. This may require a specific function to eliminate the unecessary columns.